requirements.txt

langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-milvus>=0.1.0
langchain-groq>=0.2.0
langgraph>=0.2.0
fastembed>=0.3.0
pymilvus>=2.4.0
gradio>=4.0.0
python-dotenv>=1.0.0
markdown>=3.5.0
weasyprint>=61.0.0
PyPDF2>=3.0.0
pypdf>=4.0.0